projects/wml-table-zero/src/lib/wml-table-zero-sample-row/wml-table-zero-sample-row.component.spec.ts
// testing
import { ComponentFixture } from '@angular/core/testing';
import { resetImports } from '../test-utils/mock-imports';
import { resetProviders } from '../test-utils/mock-providers';
import { wmlTestUtils } from '../test-utils/test-utils';
// rxjs
import { Subject } from 'rxjs';

import { WmlTableZeroSampleRowComponent } from './wml-table-zero-sample-row.component';


describe('WmlTableZeroSampleRowComponent', () => {
  let cpnt: WmlTableZeroSampleRowComponent;
  let fixture: ComponentFixture<WmlTableZeroSampleRowComponent>;

  beforeEach(async () => {



    await wmlTestUtils.configureTestingModuleForComponents(WmlTableZeroSampleRowComponent);


    ({fixture, cpnt} = wmlTestUtils.grabComponentInstance(WmlTableZeroSampleRowComponent));
    fixture.detectChanges()
  })

  describe("init", () => {

    it("should create", () => {
      expect(cpnt).toBeTruthy()
    })

    it("should have all values initalize properly", () => {
      expect(cpnt.myClass).toEqual('WmlTableZeroSampleRowView')
    })

    it("should have all properties be the correct class instance", () => {
      expect(cpnt.ngUnsub).toBeInstanceOf(Subject<void>)
    })
  })

  describe("ngOnDestroy",()=>{

    beforeEach(()=>{
      spyOn(cpnt.ngUnsub,'next')
      spyOn(cpnt.ngUnsub,'complete')
    })

    it(` when called |
     as appropriate |
     does the required action `,()=>{
        // act
        cpnt.ngOnDestroy();

        // assert
        expect(cpnt.ngUnsub.next).toHaveBeenCalled();
        expect(cpnt.ngUnsub.complete).toHaveBeenCalled();
    })
  })
});
